Rename feed related modules
