+++ feed_data/build.gradle
@@ -2,7 +2,7 @@ apply plugin: 'kotlin'
 apply plugin: 'kotlin-kapt'
 
 dependencies {
-    implementation project(':domain')
+    implementation project(':feed_domain')
     implementation project(':remote')
 
     implementation libraries.kotlinCoroutinesCore
