+++ feed_domain/src/main/java/com/muvi/feed_domain/Film.kt
@@ -1,3 +1,3 @@
-package com.muvi.domain
+package com.muvi.feed_domain
 
 data class Film(val id: String, val name: String)
