+++ feed_domain/src/main/java/com/muvi/feed_domain/FilmRepository.kt
@@ -1,4 +1,4 @@
-package com.muvi.domain
+package com.muvi.feed_domain
 
 interface FilmRepository {
