+++ settings.gradle
@@ -1,6 +1,6 @@
 include ':app'
-include ':data'
-include ':domain'
+include ':feed_data'
+include ':feed_domain'
 include ':feed'
 include ':film_detail'
 include ':film_detail_data'
